refactor(split): move central key handling out of Bluetooth directory
