+++ app/include/zmk/split/central.h
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2023 The ZMK Contributors
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+#pragma once
+
+#include <zmk/events/position_state_changed.h>
+
+void zmk_position_state_change_handle(struct zmk_position_state_changed *ev);
